Support computing adjusted scores
Given a set of tests that are not supposed to pass, allow computing
the fraction of the score in each focus area that corresponds to, and
the score of a run against the adjusted total.

This is implemented by allowing test results to be annotated with an
"expected" status. When a test does not pass in a run, and has an
expected annotation, and the expectation is not a passing
status (i.e. PASS or OK), then we add it as an "expected" failure.

These are returned as a vector of tuples, representing (expected
failure score, adjusted score). So, for example if a focus area has 10
tests, 7 passes and one expected failure, the score will be 70.0, the
expected failure score will be 10.0 (since 10% of the tests are
expected to fail) and the adjusted score will be 77.8 (since 7/9 tests
that are expected to pass do pass).